Build 139 has been transfered successfully

+++ models/Contact.php
@@ -584,8 +584,12 @@ class Contact extends \yii\db\ActiveRecord
 
     public static function addContInPool($contact_id, $manager_id, $tag_id, $order_token)
     {
-        $cont_pools = TempContactsPool::find(['manager_id' => $manager_id])->all();
-        foreach ($cont_pools as $cont_pool) {
+//        $cont_pools = TempContactsPool::find(['manager_id' => $manager_id])->all();
+//        foreach ($cont_pools as $cont_pool) {
+//            $cont_pool->delete();
+//        }
+        $cont_pool = TempContactsPool::findOne(['contact_id' => $contact_id, 'manager_id' => $manager_id, 'tag_id' => $tag_id]);
+        if ($cont_pool) {
             $cont_pool->delete();
         }
